requirements.txt
flet>=0.10.0
pymongo>=4.5.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
gridfs>=4.5.0